refactor done, implemented routes
