+++ client/package.json
@@ -7,6 +7,7 @@
     "bulma": "^0.7.1",
     "react": "^16.4.2",
     "react-dom": "^16.4.2",
+    "react-router-dom": "^5.0.1",
     "react-scripts": "1.1.5",
     "sweetalert2": "^7.26.12"
   },
